requirements/base.txt
psycopg2==2.8.6 --no-binary psycopg2
factory-boy==3.2.0
sqlalchemy-migrate
sqlalchemy==1.4.0
selenium
beautifulsoup4==4.9.3
PyYAML==5.4.1
